Extremes for Julia documentation

High level API

Extremes.getcluster — Function

getcluster(y::Array{<:Real,1}, u₁::Real , u₂::Real=0)

Returns a DataFrame with clusters for exceedance models. A cluster is defined as a sequence where values are higher than u₂ with at least a value higher than threshold u₁.

Extremes.getcluster — Function

getcluster(df::DataFrame, u₁::Real, u₂::Real=0)

Returns a DataFrame with clusters for exceedance models. A cluster is defined as a sequence where values are higher than u₂ with at least a value higher than threshold u₁.

Extremes.gevfitbayes — Method

gevfitbayes(y::Array{<:Real}; warmup::Int=0, niter::Int=1000, thin::Int=1, stepSize::Array{<:Real,1}=[.1,.1,.05])

Fits a GEV...

Extremes.gevfit — Method

gevfit(y::Vector{<:Real},initialvalues::Vector{<:Real}=Float64[])

Fit the Generalized Extreme Value (GEV) distribution to data using maximum likelihood. Return an object of Type GeneralizedExtremeValue.

gevfit(y): Fits the GEV distribution to the data y.

gevfit(y, initialvalues=initialvalues): Fits the GEV distribution to the data y using initialvalues as initial values.

Missing docstring.

Missing docstring for gevfit(y::Array{Float64,1}, location_covariate::Array{Float64,1}; initialvalues::Array{Float64,1}=Float64[]). Check Documenter's build log for details.

Missing docstring.

Missing docstring for gpdfit(y::Array{T} where T<:Real; threshold::Real=0.0). Check Documenter's build log for details.

Extremes.gpdfitbayes — Method

gpdfitbayes(data::Array{Float64,1}; threshold::Real=0, niter::Int = 10000, warmup::Int = 5000,  thin::Int = 1, stepSize::Array{<:Real,1}=[.1,.1])

Fits a Generalized Pareto Distribution (GPD)

Low level API

Extremes.gumbelfitpwmom — Method

gumbelfitpwmom(x::Array{T,1} where T<:Real)

Fits a Gumbel distribution using ...

Extremes.gevfitlmom — Method

gevfitlmom(x::Array{T,1} where T<:Real)

Fit a GEV distribution with L-Moment method.

Missing docstring.

Missing docstring for Extremes.getinitialvalues(y::Array{T,1} where T<:Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Extremes.gevhessian(y::Array{N,1} where N<:Real,μ::Real,σ::Real,ξ::Real). Check Documenter's build log for details.

Extremes.gpdfitmom — Method

gpdfitmom(y::Array{T} where T<:Real; threshold::Real=0.0)

Fit a Generalized Pareto Distribution over y.