Extremes for Julia documentation

High level API

Missing docstring.

Missing docstring for getcluster(y::Array{<:Real,1}, u₁::Real , u₂::Real=0.0). Check Documenter's build log for details.

Missing docstring.

Missing docstring for getcluster(df::DataFrame, u₁::Real, u₂::Real=0.0). Check Documenter's build log for details.

Missing docstring.

Missing docstring for gevfitbayes(y::Array{<:Real}; warmup::Int=0, niter::Int=1000, thin::Int=1, stepSize::Array{<:Real,1}=[.1,.1,.05]). Check Documenter's build log for details.

Extremes.gevfit — Method

gevfit(y::Vector{<:Real};
    locationcov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}(),
    logscalecov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}(),
    shapecov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}())

Fit the Generalized Extreme Value (GEV) distribution by maximum likelihood to the vector of data y.

The optional parameter locationcov is a vector containing the covariates for the parameter μ. The optional parameter logscalecov is a vector containing the covariates for the parameter σ. The optional parameter shapecov is a vector containing the covariates for the parameter ξ.

Example with a non-stationary location parameter:

using Extremes, Distributions

# Sample size
n = 300

# Covariate
x = collect(1:n)

# Location as function of the covariate
μ = x*1/100

# Sample from the non-stationary GEV distribution
pd = GeneralizedExtremeValue.(μ,1,.1)
y = rand.(pd)

# Estimate the parameters
gevfit(y, locationcov = [Variable("x", x)])

The covariate may be standardized to facilitate the estimation.

Extremes.gevfit — Method

gevfit(y::Vector{<:Real}, initialvalues::Vector{<:Real};
    locationcov::Vector{<:DataItem} = Vector{Variable}(),
    logscalecov::Vector{<:DataItem} = Vector{Variable}(),
    shapecov::Vector{<:DataItem} = Vector{Variable}(),)::MaximumLikelihoodEVA

Fit the Generalized Extreme Value (GEV) distribution by maximum likelihood to the vector of data y using the intial values initialvalues.

The covariate may be standardized to facilitate the estimation.

Missing docstring.

Missing docstring for gpdfit(y::Array{T} where T<:Real; threshold::Real=0.0). Check Documenter's build log for details.

Missing docstring.

Missing docstring for gpdfitbayes(data::Array{Float64,1}; threshold::Real=0, niter::Int = 10000, warmup::Int = 5000, thin::Int = 1, stepSize::Array{<:Real,1}=[.1,.1]). Check Documenter's build log for details.

Low level API

Missing docstring.

Missing docstring for Extremes.gumbelfitpwmom(x::Array{T,1} where T<:Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Extremes.gevfitlmom(x::Array{T,1} where T<:Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Extremes.getinitialvalues(y::Array{T,1} where T<:Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Extremes.gevhessian(y::Array{N,1} where N<:Real,μ::Real,σ::Real,ξ::Real). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Extremes.gpdfitmom(y::Array{T} where T<:Real; threshold::Real=0.0). Check Documenter's build log for details.