Functions

Extremes.cint — Function

cint(rl::ReturnLevel{BayesianEVA{BlockMaxima}};, confidencelevel::Real=.95)::Vector{Vector{Real}}

Compute the confidence interval for the return level corresponding to the return period returnPeriod from the fitted model fm with confidence level confidencelevel.

Extremes.cint — Function

cint(rl::ReturnLevel{MaximumLikelihoodEVA{ThresholdExceedance}}, threshold::Real, nobservation::Int,
    nobsperblock::Int, confidencelevel::Real=.95)::Vector{Vector{Real}}

Compute the confidence intervel for the return level corresponding to the return period returnPeriod from the fitted model fm with confidence level confidencelevel.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.cint — Function

cint(fm::MaximumLikelihoodEVA, clevel::Real=.95)::Array{Array{Float64,1},1}

Compute the Wald parameter confidence intervals using the approximate parameter estimates covariance matrix.

Extremes.cint — Function

cint(fm::pwmEVA, clevel::Real=.95, nboot::Int=1000)::Array{Array{Float64,1},1}

Estimate the parameter estimates highest posterior density interval.

Extremes.cint — Function

cint(rl::ReturnLevel{pwmEVA{BlockMaxima, T}} where T<:Distribution, confidencelevel::Real=.95)::Vector{Vector{Real}}

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

Extremes.cint — Function

cint(rl::ReturnLevel{BayesianEVA{ThresholdExceedance}}, threshold::Real, nobservation::Int,
    nobsperblock::Int, confidencelevel::Real=.95)::Vector{Vector{Real}}

Compute the confidence interval for the return level corresponding to the return period returnPeriod from the fitted model fm with the confidence level confidencelevel.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.cint — Function

cint(rl::ReturnLevel{MaximumLikelihoodEVA{BlockMaxima}}, confidencelevel::Real=.95)::Vector{Vector{Real}}

Compute the confidence intervel for the return level corresponding to the return period returnPeriod from the fitted model fm with confidence level confidencelevel.

Extremes.cint — Function

cint(rl::ReturnLevel{pwmEVA{ThresholdExceedance, T}} where T<:Distribution, threshold::Real, nobservation::Int, nobsperblock::Int, confidencelevel::Real=.95)::Vector{Vector{Real}}

Compute the confidence intervel for the return level corresponding to the return period returnPeriod from the fitted model fm with confidence level confidencelevel.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.cint — Function

cint(fm::pwmEVA, clevel::Real=.95, nboot::Int=1000)::Array{Array{Float64,1},1}

Estimate the parameter estimates confidence interval by bootstrap.

Extremes.diagnosticplots — Method

TODO : desc

Extremes.getcluster — Method

getcluster(y::Vector{<:Real}, u::Real; runlength::Int=1)::Vector{Cluster}

Threshold exceedances separated by fewer than r non-exceedances belong to the same cluster. The value r is corresponds to the runlength parameter. This approach is referred to as the runs declustering scheme (see Coles, 2001 sec. 5.3.2).

Extremes.getcluster — Method

getcluster(y::Vector{<:Real}, u₁::Real, u₂::Real)::Vector{Cluster}

Returns a DataFrame with clusters for exceedance models. A cluster is defined as a sequence where values are higher than u₂ with at least a value higher than threshold u₁.

Extremes.gevfit — Method

gevfit(y::Vector{<:Real}, initialvalues::Vector{<:Real};
    locationcov::Vector{<:DataItem} = Vector{Variable}(),
    logscalecov::Vector{<:DataItem} = Vector{Variable}(),
    shapecov::Vector{<:DataItem} = Vector{Variable}(),)::MaximumLikelihoodEVA

Fit the Generalized Extreme Value (GEV) distribution by maximum likelihood to the vector of data y using the intial values initialvalues.

The covariate may be standardized to facilitate the estimation.

Extremes.gevfit — Method

gevfit(y::Vector{<:Real};
    locationcov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}(),
    logscalecov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}(),
    shapecov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}())

Fit the Generalized Extreme Value (GEV) distribution by maximum likelihood to the vector of data y.

The optional parameter locationcov is a vector containing the covariates for the parameter μ. The optional parameter logscalecov is a vector containing the covariates for the parameter σ. The optional parameter shapecov is a vector containing the covariates for the parameter ξ.

Example with a non-stationary location parameter:

using Extremes, Distributions

# Sample size
n = 300

# Covariate
x = collect(1:n)

# Location as function of the covariate
μ = x*1/100

# Sample from the non-stationary GEV distribution
pd = GeneralizedExtremeValue.(μ,1,.1)
y = rand.(pd)

# Estimate the parameters
gevfit(y, locationcov = [Variable("x", x)])

The covariate may be standardized to facilitate the estimation.

Extremes.gevfit — Method

gevfit(model::BlockMaxima, initialvalues::Vector{<:Real})::MaximumLikelihoodEVA

Fit the Generalized Extreme Value (GEV) distribution by maximum likelihood of the BlockMaxima model model.

Extremes.gevfit — Method

gevfit(df::DataFrame, datacol::Symbol, initialvalues::Vector{<:Real};
    locationcovid::Vector{Symbol}=Symbol[],
    logscalecovid::Vector{Symbol}=Symbol[],
    shapecovid::Vector{Symbol}=Symbol[])::MaximumLikelihoodEVA

Fit the Generalized Extreme Value (GEV) distribution by maximum likelihood to the vector of data contained in the dataframe df at the column datacol using the initial values ìnitialvalues.

Extremes.gevfit — Method

gevfit(df::DataFrame, datacol::Symbol;
    locationcovid::Vector{Symbol}=Symbol[],
    logscalecovid::Vector{Symbol}=Symbol[],
    shapecovid::Vector{Symbol}=Symbol[])::MaximumLikelihoodEVA

Fit the Generalized Extreme Value (GEV) distribution by maximum likelihood to the vector of data contained in the dataframe df at the column datacol.

Extremes.gevfitbayes — Method

gevfitbayes(y::Vector{<:Real};
    locationcov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}(),
    logscalecov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}(),
    shapecov::Vector{Vector{T}} where T<:Real = Vector{Vector{Float64}}(),
    niter::Int=5000, warmup::Int=2000)

Fit the Generalized Extreme Value (GEV) distribution under the Bayesian paradigm to the vector of data y.

The optional parameter locationcov is a vector containing the covariates for the parameter μ. The optional parameter logscalecov is a vector containing the covariates for the parameter σ. The optional parameter shapecov is a vector containing the covariates for the parameter ξ.

The covariate may be standardized to facilitate the estimation.

A random sample of the posterior distribution is generated using the NUTS algortihm.

Only flat prior is now supported.

Example with a non-stationary location parameter:

# Sample size
n = 300

# Covariate
x = collect(1:n)

# Location as function of the covariate
μ = x*1/100

# Sample from the non-stationary GEV distribution
pd = GeneralizedExtremeValue.(μ,1,.1)
y = rand.(pd)

# Estimate the parameters
gevfitbayes(y, locationcov = [Variable("x", x)])

Extremes.gevfitbayes — Method

gevfitbayes(model::BlockMaxima; niter::Int=5000, warmup::Int=2000)

Fit a non-stationary Generalized Extreme Value (GEV) distribution under the Bayesian paradigm of the BlockMaxima model model.

Extremes.gevfitbayes — Method

gevfitbayes(df::DataFrame, datacol::Symbol;
    locationcovid::Vector{Symbol}=Symbol[],
    logscalecovid::Vector{Symbol}=Symbol[],
    shapecovid::Vector{Symbol}=Symbol[],
    niter::Int=5000, warmup::Int=2000)::MaximumLikelihoodEVA

Fit a Generalized Extreme Value (GEV) distribution under the Bayesian paradigm to the vector of data contained in the dataframe df at the column datacol.

Extremes.gevfitpwm — Method

gevfitpwm(y::Vector{<:Real})::pwmEVA

Estimate the Generalized Extreme value distribution parameters with the probability weighted moments as described in Hosking et al. (1985).

Reference: Hosking, J. R. M., Wallis, J. R. and Wood, E. F. (1985). Estimation of the generalized extreme-value distribution by the method of probability-weighted moments. Technometrics, 27, 251-261.

Extremes.gevfitpwm — Method

gevfitpwm(model::BlockMaxima)::pwmEVA

Estimate the Generalized Extreme value distribution parameters with the probability weighted moments as described in Hosking et al. (1985).

With the methods of moments, it is not possible to include covariates in the model. If covariates are provided, they are ignored and the stationary model is fitted.

Reference: Hosking, J. R. M., Wallis, J. R. and Wood, E. F. (1985). Estimation of the generalized extreme-value distribution by the method of probability-weighted moments. Technometrics, 27, 251-261.

Extremes.gevfitpwm — Method

gevfitpwm(df::DataFrame, datacol::Symbol)::pwmEVA

Estimate the Generalized Extreme value distribution parameters with the probability weighted moments as described in Hosking et al. (1985).

Reference: Hosking, J. R. M., Wallis, J. R. and Wood, E. F. (1985). Estimation of the generalized extreme-value distribution by the method of probability-weighted moments. Technometrics, 27, 251-261.

Extremes.gpfit — Method

gpfit(y::Vector{<:Real}, initialvalues::Vector{<:Real};
    logscalecov::Vector{<:DataItem} = Vector{Variable}(),
    shapecov::Vector{<:DataItem} = Vector{Variable}())::MaximumLikelihoodEVA

Fit the Generalized Pareto (GP) distribution by maximum likelihood to the vector of data y using the initial values initialvalues.

The optional parameter logscalecov is a vector containing the covariates for the parameter σ. The optional parameter shapecov is a vector containing the covariates for the parameter ξ.

The covariate may be standardized to facilitate the estimation.

Extremes.gpfit — Method

gpfit(y::Vector{<:Real};
    logscalecov::Vector{<:DataItem} = Vector{Variable}(),
    shapecov::Vector{<:DataItem} = Vector{Variable}())::MaximumLikelihoodEVA

Fit the Generalized Pareto (GP) distribution by maximum likelihood to the vector of data y.

The optional parameter logscalecov is a vector containing the covariates for the parameter σ. The optional parameter shapecov is a vector containing the covariates for the parameter ξ.

Example with a non-stationary location parameter:

# Sample size
n = 300

# Covariate
x = collect(1:n)

# Location as function of the covariate
ϕ = x*1/500
σ = exp.(ϕ)

# Sample from the non-stationary GEV distribution
pd = GeneralizedPareto.(σ,.1)
y = rand.(pd)

# Estimate the parameters
gpfit(y, logscalecov = [Variable("x", x)])

The covariate may be standardized to facilitate the estimation.

Extremes.gpfit — Method

gpfit(df::DataFrame, datacol::Symbol, initialvalues::Vector{<:Real};
    logscalecovid::Vector{Symbol}=Symbol[],
    shapecovid::Vector{Symbol}=Symbol[])::MaximumLikelihoodEVA

Fit a Generalized Pareto (GP) distribution by maximum likelihood to the vector of data contained in the dataframe df at the column datacol using the initial values initialvalues.

Extremes.gpfit — Method

gpfit(df::DataFrame, datacol::Symbol;
    logscalecovid::Vector{Symbol}=Symbol[],
    shapecovid::Vector{Symbol}=Symbol[])::MaximumLikelihoodEVA

Fit a Generalized Pareto (GP) distribution by maximum likelihood to the vector of data contained in the dataframe df at the column datacol.

Extremes.gpfit — Method

gpfit(model::ThresholdExceedance, initialvalues::Vector{<:Real})::MaximumLikelihoodEVA

Fit the Generalized Pareto (GP) distribution by maximum likelihood to the ThresholdExceedance model.

Extremes.gpfitbayes — Method

gpfitbayes(y::Vector{<:Real};
     logscalecov::Vector{<:DataItem} = Vector{Variable}(),
     shapecov::Vector{<:DataItem} = Vector{Variable}(),
     niter::Int=5000, warmup::Int=2000)::BayesianEVA

Fit a non-stationary Generalized Pareto (GEV) distribution under the Bayesian paradigm to the vector of data contained in the Vector y.

The optional parameter logscalecov is a vector containing the covariates for the parameter σ. The optional parameter shapecov is a vector containing the covariates for the parameter ξ.

The covariate may be standardized to facilitate the estimation.

A random sample of the posterior distribution is generated using the NUTS algortihm.

Only flat prior is now supported.

Example with a non-stationary location parameter:

# Sample size
n = 300

# Covariate
x = collect(1:n)

# Location as function of the covariate
ϕ = x*1/500
σ = exp.(ϕ)

# Sample from the non-stationary GEV distribution
pd = GeneralizedPareto.(σ,.1)
y = rand.(pd)

# Estimate the parameters
gpfitbayes(y, logscalecov = [Variable("x", x)])

Extremes.gpfitbayes — Method

gpfitbayes(df::DataFrame, datacol::Symbol;
    logscalecovid::Vector{Symbol}=Symbol[],
    shapecovid::Vector{Symbol}=Symbol[],
    niter::Int=5000, warmup::Int=2000)::MaximumLikelihoodEVA

Fit a Generalized Pareto (GP) distribution under the Bayesian paradigm to the vector of data contained in the dataframe df at the column datacol.

Extremes.gpfitbayes — Method

gpfitbayes(model::ThresholdExceedance, niter::Int=5000, warmup::Int=2000)::BayesianEVA

Fit the Generalized Pareto (GP) distribution under the Bayesian paradigm to the ThresholdExceedance model.

A random sample from the posterior distribution is generated using the NUTS algortihm.

Only flat prior is now supported.

Extremes.gpfitpwm — Method

gpfitpwm(y::Vector{<:Real})::pwmEVA

Estimate the Generalized Pareto distribution parameters with the probability weighted moments as described in Hosking & Wallis (1987).

Reference: Hosking, J. R. M. and Wallis, J. R. (1987). Parameter and Quantile Estimation for the Generalized Pareto Distribution, Technometrics, 29(3), 339-349.

Extremes.gpfitpwm — Method

gpfitpwm(df::DataFrame, datacol::Symbol)::pwmEVA

Estimate the Generalized Pareto distribution parameters with the probability weighted moments as described in Hosking & Wallis (1987).

Reference: Hosking, J. R. M. and Wallis, J. R. (1987). Parameter and Quantile Estimation for the Generalized Pareto Distribution, Technometrics, 29(3), 339-349.

Extremes.gpfitpwm — Method

gpfitpwm(model::ThresholdExceedance)::pwmEVA

Estimate the Generalized Pareto distribution parameters with the probability weighted moments as described in Hosking & Wallis (1987).

Reference: Hosking, J. R. M. and Wallis, J. R. (1987). Parameter and Quantile Estimation for the Generalized Pareto Distribution, Technometrics, 29(3), 339-349.

Extremes.histplot — Method

TODO : desc

Extremes.histplot_data — Method

TODO : desc

Extremes.histplot_data — Method

TODO : desc

Extremes.histplot_data — Method

TODO : desc

Extremes.load — Method

load(name::String)::DataFrame

Returns the data associated with the name.

Extremes.mrlplot — Function

mrlplot(y::Vector{<:Real}, steps::Int = 100)::Plot

Show the mean residual life from vector y using the set of thresholds from the minimum(y) value to the second-to-last larger value in steps steps.

Extremes.mrlplot_data — Function

mrlplot_data(y::Vector{<:Real}, steps::Int = 100)::DataFrame

Compute the mean residual life from vector y using the set of thresholds from the minimum(y) value to the second-to-last larger value in steps steps.

Extremes.parametervar — Function

parametervar(fm::pwmEVA, nboot::Int=1000)

Estimate the parameter estimates covariance matrix by bootstrap.

Extremes.parametervar — Method

parametervar(fm::BayesianEVA)::Array{Float64, 2}

Compute the covariance parameters estimate of the fitted model fm.

Extremes.parametervar — Method

parametervar(fm::MaximumLikelihoodEVA)::Array{Float64, 2}

Compute the covariance parameters estimate of the fitted model fm.

Extremes.probplot — Method

TODO : desc

Extremes.probplot_data — Method

TODO : desc

Extremes.probplot_data — Method

TODO : desc

Extremes.probplot_data — Method

TODO : desc

Extremes.probplot_std_data — Method

TODO : desc

Extremes.qqplot — Method

TODO : desc

Extremes.qqplot_data — Method

TODO : desc

Extremes.qqplot_data — Method

TODO : desc

Extremes.qqplot_data — Method

TODO : desc

Extremes.qqplot_std_data — Method

TODO : desc

Extremes.returnlevel — Method

returnlevel(fm::BayesianEVA{BlockMaxima}, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

Extremes.returnlevel — Method

returnlevel(fm::BayesianEVA{ThresholdExceedance}, threshold::Real, nobservation::Int,
    nobsperblock::Int, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.returnlevel — Method

returnlevel(fm::MaximumLikelihoodEVA{BlockMaxima}, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

Extremes.returnlevel — Method

returnlevel(fm::MaximumLikelihoodEVA{ThresholdExceedance}, threshold::Real, nobservation::Int,
    nobsperblock::Int, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.returnlevel — Method

returnlevel(fm::pwmEVA{BlockMaxima, T} where T<:Distribution, returnPeriod::Real)::ReturnLevel

Compute the confidence intervel for the return level corresponding to the return period returnPeriod from the fitted model fm with confidence level confidencelevel.

Extremes.returnlevel — Method

returnlevel(fm::pwmEVA{ThresholdExceedance, T} where T<:Distribution, threshold::Real, nobservation::Int,
    nobsperblock::Int, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.returnlevelplot — Method

TODO : desc

Extremes.returnlevelplot_data — Method

TODO : desc

Extremes.returnlevelplot_data — Method

TODO : desc

Extremes.returnlevelplot_data — Method

TODO : desc

Types

Extremes.BlockMaxima — Method

BlockMaxima(data::Vector{<:Real};
    locationcov::Vector{Variable} = Vector{Variable}(),
    logscalecov::Vector{Variable} = Vector{Variable}(),
    shapecov::Vector{Variable} = Vector{Variable}())::BlockMaxima

Creates a BlockMaxima structure.

Extremes.ThresholdExceedance — Method

ThresholdExceedance(exceedances::Vector{<:Real};
    logscalecov::Vector{<:DataItem} = Vector{Variable}(),
    shapecov::Vector{<:DataItem} = Vector{Variable}())::ThresholdExceedance

Creates a ThresholdExceedance structure.

Extremes.VariableStd — Method

VariableStd(name::String, z::Vector{<:Real})::VariableStd

Construct a VariableStd type from the standardized vector z with the name name.