Functions

Parameter estimation

Extremes.gevfit — Function

gevfit()

Estimate the GEV parameters by maximum likelihood.

Implementation

The function uses Nelder-Mead solver implemented in the Optim.jl package to find the point where the log-likelihood is maximal.

The GEV parameters can be modeled as function of covariates as follows:

\[μ = X₁ × β₁,\]

\[ϕ = X₂ × β₂,\]

\[ξ = X₃ × β₃.\]

The covariates are standardized before estimating the parameters to help fit the model. They are transformed back on their original scales before returning the fitted model.

See also gevfit for the other methods, gevfitpwm, gevfitbayes and BlockMaxima.

Extremes.gevfit — Method

gevfit(y,
    initialvalues,
    locationcov = Vector{Variable}(),
    logscalecov = Vector{Variable}(),
    shapecov = Vector{Variable}()
    )

Estimate the GEV parameters.

Arguments

y::Vector{<:Real}: the vector of block maxima.

initialvalues::Vector{<:Real}: Vector of parameters initial values.

locationcov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the location parameter.

logscalecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the log-scale parameter.

shapecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the shape parameter.

Extremes.gevfit — Method

gevfit(y,
    locationcov = Vector{Variable}(),
    logscalecov = Vector{Variable}(),
    shapecov = Vector{Variable}()
    )

Estimate the GEV parameters.

Arguments

y::Vector{<:Real}: The vector of block maxima.

locationcov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the location parameter.

logscalecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the log-scale parameter.

shapecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the shape parameter.

Extremes.gevfit — Method

gevfit(model::BlockMaxima, initialvalues::Vector{<:Real})

Estimate the parameters of the BlockMaxima model using the given initialvalues.

Extremes.gevfit — Method

gevfit(df::DataFrame,
    datacol::Symbol,
    locationcovid = Vector{Symbol}(),
    logscalecovid = Vector{Symbol}(),
    shapecovid = Vector{Symbol}()
    )

Estimate the GEV parameters.

Arguments

df::DataFrame: The dataframe containing the data.

datacol::Symbol: The symbol of the column of df containing the block maxima data.

initialvalues::Vector{<:Real}: Vector of parameters initial values.

locationcovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the location parameter.

logscalecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the log-scale parameter.

shapecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the shape parameter.

Extremes.gevfit — Method

gevfit(df::DataFrame,
    datacol::Symbol,
    locationcovid = Vector{Symbol}(),
    logscalecovid = Vector{Symbol}(),
    shapecovid = Vector{Symbol}()
    )

Estimate the GEV parameters.

Arguments

df::DataFrame: The dataframe containing the data.

datacol::Symbol: The symbol of the column of df containing the block maxima data.

locationcovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the location parameter.

logscalecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the log-scale parameter.

shapecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the shape parameter.

Extremes.gpfit — Function

gpfit(...)

Estimate the GP parameters by maximum likelihood.

Data provided must be the exceedances above the threshold, i.e. the data above the threshold minus the threshold.

Implementation

The function uses Nelder-Mead solver implemented in the Optim.jl package to find the point where the log-likelihood is maximal.

The GP parameters can be modeled as function of covariates as follows:

\[ϕ = X₂ × β₂,\]

\[ξ = X₃ × β₃.\]

The covariates are standardized before estimating the parameters to help fit the model. They are transformed back on their original scales before returning the fitted model.

See also gpfit for the other methods, gpfitpwm, gpfitbayes and ThresholdExceedance.

Extremes.gpfit — Method

gpfit(y,
    initialvalues;
    logscalecov = Vector{Variable}(),
    shapecov = Vector{Variable}()
    )

Estimate the GP parameters

Arguments

y::Vector{<:Real}: The vector of exceedances.

initialvalues::Vector{<:Real}: The vector of parameters initial values.

logscalecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the log-scale parameter.

shapecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the shape parameter.

Extremes.gpfit — Method

gpfit(y,
    logscalecov = Vector{Variable}(),
    shapecov = Vector{Variable}()
    )

Estimate the GP parameters

Arguments

y::Vector{<:Real}: The vector of exceedances.

logscalecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the log-scale parameter.

shapecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the shape parameter.

Extremes.gpfit — Method

gpfit(df::DataFrame,
    datacol::Symbol,
    logscalecovid = Vector{Symbol}(),
    shapecovid = Vector{Symbol}()
    )

Estimate the GP parameters

Arguments

df::DataFrame: The dataframe containing the data.

datacol::Symbol: The symbol of the column of df containing the exceedances.

initialvalues::Vector{<:Real}: Vector of parameters initial values.

logscalecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the log-scale parameter.

shapecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the shape parameter.

Extremes.gpfit — Method

gpfit(df::DataFrame,
    datacol::Symbol,
    logscalecovid = Vector{Symbol}(),
    shapecovid = Vector{Symbol}()
    )

Estimate the GP parameters

Arguments

df::DataFrame: The dataframe containing the data.

datacol::Symbol: The symbol of the column of df containing the exceedances.

logscalecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the log-scale parameter.

shapecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the shape parameter.

Extremes.gpfit — Method

gpfit(model::ThresholdExceedance, initialvalues::Vector{<:Real})

Estimate the parameters of the ThresholdExceedance model using the given initialvalues.

Extremes.gevfitpwm — Function

gevfitpwm(...)

Estimate the GEV parameters with the probability weighted moments.

Implementation

Estimation with the probability weighted moments, as described by Hosking et al. (1985), is only possible in the stationary case.

See also gevfitpwm for the other methods, gevfit, gevfitbayes and BlockMaxima.

Reference

Hosking, J. R. M., Wallis, J. R. and Wood, E. F. (1985). Estimation of the generalized extreme-value distribution by the method of probability-weighted moments. Technometrics, 27:251-261.

Extremes.gevfitpwm — Method

gevfitpwm(y::Vector{<:Real})

Estimate the GEV parameters with the probability weighted moments.

Extremes.gevfitpwm — Method

gevfitpwm(model::BlockMaxima)

Estimate the GEV parameters with the probability weighted moments.

Extremes.gevfitpwm — Method

gevfitpwm(df::DataFrame, datacol::Symbol)

Estimate the GEV parameters with the probability weighted moments.

Block maxima data are in the column datacol of the dataframe df.

Extremes.gpfitpwm — Function

gpfitpwm(...)

Estimate the GP parameters with the probability weighted moments.

Implementation

Estimation with the probability weighted moments, as described by Hosking and Wallis (1987), is only possible in the stationary case.

See also gpfitpwm for the other methods, gpfit, gpfitbayes and ThresholdExceedance.

Reference

Hosking, J. R. M. and Wallis, J. R. (1987). Parameter and quantile estimation for the Generalized Pareto distribution, Technometrics, 29:339-349.

Extremes.gpfitpwm — Method

gpfitpwm(y::Vector{<:Real})

Estimate the GP parameters with the probability weighted moments.

Extremes.gpfitpwm — Method

gpfitpwm(df::DataFrame, datacol::Symbol)

Estimate the GP parameters with the probability weighted moments.

Block maxima data are in the column datacol of the dataframe df.

Extremes.gpfitpwm — Method

gpfitpwm(model::ThresholdExceedance)

Estimate the GP parameters with the probability weighted moments.

Extremes.gevfitbayes — Function

gevfitbayes(..., niter::Int=5000, warmup::Int=2000)

Generate a sample from the GEV parameters' posterior distribution.

Arguments

niter::Int = 5000: The total number of MCMC iterations.

warmup::Int = 2000: The number of warmup iterations (burn-in).

Implementation

The function uses the No-U-Turn Sampler (NUTS; Hoffman and Gelman, 2014) implemented in the Mamba.jl package to generate a random sample from the posterior distribution.

Currently, only the improper uniform prior is implemented, i.e.

\[f_{(β₁,β₂,β₃)}(β₁,β₂,β₃) ∝ 1,\]

where

\[μ = X₁ × β₁,\]

\[ϕ = X₂ × β₂,\]

\[ξ = X₃ × β₃.\]

In the stationary case, this improper prior yields to a proper posterior if the sample size is larger than 3 (Northrop and Attalides, 2016).

The covariates are standardized before estimating the parameters to help fit the model. They are transformed back on their original scales before returning the fitted model.

See also gevfitbayes for the other methods, gevfitpwm, gevfit and BlockMaxima.

References

Hoffman M. D. and Gelman A. (2014). The No-U-Turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo. Journal of Machine Learning Research, 15:1593–1623.

Paul J. Northrop P. J. and Attalides N. (2016). Posterior propriety in Bayesian extreme value analyses using reference priors. Statistica Sinica, 26:721-743.

Extremes.gevfitbayes — Method

gevfitbayes(y,
    locationcov = Vector{Variable}(),
    logscalecov = Vector{Variable}(),
    shapecov = Vector{Variable}(),
    niter::Int=5000,
    warmup::Int=2000
    )

Generate a sample from the GEV parameters' posterior distribution.

Arguments

y::Vector{<:Real}: The vector of block maxima.

locationcov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the location parameter.

logscalecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the log-scale parameter.

shapecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the shape parameter.

Extremes.gevfitbayes — Method

gevfitbayes(model::BlockMaxima;
    niter::Int=5000,
    warmup::Int=2000)

Generate a sample from the BlockMaxima model parameters' posterior distribution.

Extremes.gevfitbayes — Method

gevfitbayes(df::DataFrame,
    datacol::Symbol,
    locationcovid = Vector{Symbol}(),
    logscalecovid = Vector{Symbol}(),
    shapecovid = Vector{Symbol}(),
    niter::Int=5000,
    warmup::Int=2000)

Generate a sample from the GEV parameters' posterior distribution.

Arguments

df::DataFrame: The dataframe containing the data.

datacol::Symbol: The symbol of the column of df containing the block maxima data.

locationcovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the location parameter.

logscalecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the log-scale parameter.

shapecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the shape parameter.

Extremes.gpfitbayes — Function

gpfitbayes(..., niter::Int=5000, warmup::Int=2000)

Generate a sample from the GP parameters' posterior distribution.

Data provided must be the exceedances above the threshold, i.e. the data above the threshold minus the threshold.

Arguments

niter::Int = 5000: The total number of MCMC iterations

warmup::Int = 2000: The number of warmup iterations (burn-in).

Implementation

The function uses the No-U-Turn Sampler (NUTS; Hoffman and Gelman, 2014) implemented in the Mamba.jl package to generate a random sample from the posterior distribution.

Currently, only the improper uniform prior is implemented, i.e.

\[f_{(β₂,β₃)}(β₂,β₃) ∝ 1,\]

where

\[ϕ = X₂ × β₂,\]

\[ξ = X₃ × β₃.\]

In the stationary case, this improper prior yields to a proper posterior if the sample size is larger than 2 (Northrop and Attalides, 2016).

The covariates are standardized before estimating the parameters to help fit the model. They are transformed back on their original scales before returning the fitted model.

See also gpfitbayes for the other methods, gpfitpwm, gpfit and ThresholdExceedance.

Reference

Hoffman M. D. and Gelman A. (2014). The No-U-Turn sampler: adaptively setting path lengths in Hamiltonian Monte Carlo. Journal of Machine Learning Research, 15:1593–1623.

Paul J. Northrop P. J. and Attalides N. (2016). Posterior propriety in Bayesian extreme value analyses using reference priors. Statistica Sinica, 26:721-743.

Extremes.gpfitbayes — Method

gpfitbayes(y,
    logscalecov = Vector{Variable}(),
    shapecov = Vector{Variable}(),
    niter::Int=5000,
    warmup::Int=2000
    )

Generate a sample from the GP parameters' posterior distribution.

Arguments

y::Vector{<:Real}: The vector of exceedances.

logscalecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the log-scale parameter.

shapecov::Vector{<:DataItem} = Vector{Variable}(): The covariates of the shape parameter.

Extremes.gpfitbayes — Method

gpfitbayes(df::DataFrame,
    datacol::Symbol,
    logscalecovid = Vector{Symbol}(),
    shapecovid = Vector{Symbol}(),
    niter::Int=5000,
    warmup::Int=2000)

Generate a sample from the GP parameters' posterior distribution.

Arguments

df::DataFrame: The dataframe containing the data.

datacol::Symbol: The symbol of the column of df containing the exceedances.

logscalecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the log-scale parameter.

shapecovid::Vector{Symbol} = Vector{Symbol}(): The symbols of the columns of df containing the covariates of the shape parameter.

Extremes.gpfitbayes — Method

gpfitbayes(model::ThresholdExceedance;
    niter::Int=5000,
    warmup::Int=2000)

Generate a sample from the GP parameters' posterior distribution.

Methods on fitted models

Extremes.getcluster — Method

getcluster(y::Vector{<:Real}, u::Real; runlength::Int=1)::Vector{Cluster}

Extract the clusters from vector y.

Threshold exceedances separated by fewer than r non-exceedances belong to the same cluster. The value r is corresponds to the runlength parameter. This approach is referred to as the runs declustering scheme (see Coles, 2001 sec. 5.3.2).

See also Cluster.

Extremes.getcluster — Method

getcluster(y::Vector{<:Real}, u₁::Real, u₂::Real)

Extract the clusters from vector y.

A cluster is defined as a sequence of values higher than threshold u₂ with at least a value higher than threshold u₁.

See also Cluster.

Extremes.cint — Function

cint(..., confidencelevel::Real=.95)

Compute confidence interval or credible interval in the case of Bayesian estimation.

The function can be applied on any fittedEVA subtype to obtain a confidence interval on the model parameters. It can also be applied on ReturnLevel type to obtain a confidence interval on the return level.

Implementation

The method used for computing the interval depends on the estimation method. In the case of maximum likelihood estimation, the confidence intervals are computed using the Wald approximation based on the approximate parameter estimates covariance matrix. In the case of Bayesian estimation, the return interval is the highest posterior density estimate based on the MCMC sample. In the case of probability weighted moment estimation, the intervals are computed using a boostrap procedure.

Extremes.parametervar — Method

parametervar(fm::BayesianEVA)::Array{Float64, 2}

Compute the covariance parameters estimate of the fitted model fm.

Extremes.returnlevel — Method

returnlevel(fm::BayesianEVA{BlockMaxima}, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

Extremes.returnlevel — Method

returnlevel(fm::BayesianEVA{ThresholdExceedance}, threshold::Real, nobservation::Int,
    nobsperblock::Int, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.parametervar — Method

parametervar(fm::MaximumLikelihoodEVA)::Array{Float64, 2}

Compute the covariance parameters estimate of the fitted model fm.

Extremes.returnlevel — Method

returnlevel(fm::MaximumLikelihoodEVA{BlockMaxima}, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

Extremes.returnlevel — Method

returnlevel(fm::MaximumLikelihoodEVA{ThresholdExceedance}, threshold::Real, nobservation::Int,
    nobsperblock::Int, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

The threshold should be a scalar. A varying threshold is not yet implemented.

Extremes.parametervar — Function

parametervar(fm::pwmEVA, nboot::Int=1000)

Estimate the parameter estimates covariance matrix by bootstrap.

Extremes.returnlevel — Method

returnlevel(fm::pwmEVA{BlockMaxima, T} where T<:Distribution, returnPeriod::Real)::ReturnLevel

Compute the confidence intervel for the return level corresponding to the return period returnPeriod from the fitted model fm with confidence level confidencelevel.

Extremes.returnlevel — Method

returnlevel(fm::pwmEVA{ThresholdExceedance, T} where T<:Distribution, threshold::Real, nobservation::Int,
    nobsperblock::Int, returnPeriod::Real)::ReturnLevel

Compute the return level corresponding to the return period returnPeriod from the fitted model fm.

The threshold should be a scalar. A varying threshold is not yet implemented.

Diagnostic plots

Extremes.diagnosticplots — Method

diagnosticplots(fm::fittedEVA)

Diagnostic plots

Extremes.histplot — Method

histplot(fm::fittedEVA)

Histogram plot

Extremes.mrlplot — Function

mrlplot(y::Vector{<:Real}, steps::Int = 100)

Mean residual plot

See also mrlplot_data.

Extremes.probplot — Method

probplot(fm::fittedEVA)

Probability plot

Extremes.qqplot — Method

qqplot(fm::fittedEVA)

Quantile-Quantile plot

Extremes.returnlevelplot — Method

returnlevelplot(fm::fittedEVA)

Return level plot

Extremes.mrlplot_data — Function

mrlplot_data(y::Vector{<:Real}, steps::Int = 100)::DataFrame

Compute the mean residual life from vector y.

The set of thresholds ranges from minimum(y) to the second-to-last larger value in steps number of steps.

Extremes.probplot_data — Function

probplot_data(fm::fittedModel)

Return the probability plot data in a DataFrame.

Extremes.qqplot_data — Function

qqplot_data(fm::fittedModel)

Return the quantile-quantile plot data in a DataFrame.

Extremes.returnlevelplot_data — Function

returnlevelplot_data(fm::fittedModel)

Return the return level plot data in a DataFrame.

Types

Extremes.BlockMaxima — Method

BlockMaxima(data::Vector{<:Real};
    locationcov::Vector{Variable} = Vector{Variable}(),
    logscalecov::Vector{Variable} = Vector{Variable}(),
    shapecov::Vector{Variable} = Vector{Variable}())::BlockMaxima

Creates a BlockMaxima structure.

Extremes.Cluster — Type

Cluster(u₁::Real,u₂::Real,position::Vector{<:Int},value::Vector{<:Real})

Cluster type.

Extremes.EVA — Type

EVA

Abstract type containing the extreme value model types.

BlockMaxima

ThresholdExceedance

Extremes.ReturnLevel — Type

ReturnLevel

ReturnLevel type constructed by the function returnlevel.

Extremes.ThresholdExceedance — Method

ThresholdExceedance(exceedances::Vector{<:Real};
    logscalecov::Vector{<:DataItem} = Vector{Variable}(),
    shapecov::Vector{<:DataItem} = Vector{Variable}())::ThresholdExceedance

Creates a ThresholdExceedance structure.

Extremes.Variable — Type

Variable(name::String, value :: Vector{<:Real})

Construct a Variable type

Extremes.VariableStd — Method

VariableStd(name::String, z::Vector{<:Real})::VariableStd

Construct a VariableStd type from the standardized vector z with the name name.

Extremes.fittedEVA — Type

fittedEVA{T<:EVA}

Abstract type containing the fitted extreme value model types.

BayesianEVA

MaximumLikelihoodEVA

pwmEVA